Extract Node types into node.go
